add debug info, fix error multiple definition of tmp_dir, fix bug send filename with length 0, problem data type of getFileSize()

+++ src/common.h
@@ -24,7 +24,7 @@ extern const uint8_t OPENING_FILE_ERROR;
 
 extern FILE *stream;
 
-long getFileSize(char *filename);
+uint32_t getFileSize(char *filename);
 
 int print_error(char *mess);
